+++ sequana/gui/widgets.py
@@ -268,11 +268,14 @@ class ComboboxOptions(GeneralOption):
         super().__init__(option)
         self.choice = QW.QComboBox()
         self.choice.addItems(values)
+        self.choice.setStyleSheet("QComboBox { selection-background-color: #5964FF; }");
         self.layout.addWidget(self.choice)
+        
+
     def get_value(self):
         return self.choice.currentText()
     def set_enable(self, switch_bool):
-        self.check_box.setEnabled(switch_bool)
+        self.choice.setEnabled(switch_bool)
     def connect(self, task):
         self.choice.currentIndexChanged.connect(task)