+++ doc/Changelog.rst
@@ -3,6 +3,23 @@ Changelog
 
 .. contents::
 
+
+0.1.22 - March 2017
+----------------------
+
+* CHANGES:
+
+   - Sequanix/Sequana:
+     - config file can have the yml extension (in addition to yaml)
+     - dropdown widgets in the form based on the docstrings in the config file
+     - can import config to override default sequana config file
+     - subprocesses killed when the main pipeline is stopped
+   - Remoed ./reporting modules being moved to new modules_report framework 
+
+* BUGS:
+
+   - kraken fixes for special cases where all reads are unclassified
+
 0.1.21 - Feb 2017
 -------------------